+++ src/CellCms.Api/Program.cs
@@ -7,7 +7,10 @@ namespace CellCms.Api
     {
         public static void Main(string[] args)
         {
-            CreateHostBuilder(args).Build().Run();
+            CreateHostBuilder(args)
+                .Build()
+                .EnsureDatabaseIsMigrated()
+                .Run();
         }
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>